requirements.txt
audiomentations==0.34.1
numpy
pandas
scipy
torch
tqdm
python-dotenv
wandb
matplotlib
seaborn